@@ -23,15 +23,15 @@
     "@react-native-community/eslint-config": "0.0.3",
     "babel-eslint": "^10.0.2",
     "babel-jest": "24.8.0",
-    "eslint": "^5.6.0",
+    "eslint": "^5.16.0",
     "eslint-config-airbnb": "^17.1.1",
     "eslint-config-prettier": "^6.0.0",
     "eslint-config-react-app": "^4.0.1",
     "eslint-plugin-flowtype": "^2.50.3",
-    "eslint-plugin-import": "^2.18.0",
+    "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-react": "^7.14.2",
+    "eslint-plugin-react": "^7.14.3",
     "jest": "24.8.0",
     "metro-react-native-babel-preset": "0.54.1",
     "prettier": "^1.18.2",
